Updated Level 3 + Put underscores in functions
I removed the platform un level 3 because I felt it would probably be better to  introduce platforms in another level instead. The underscores I put in were to make the engine stop whining when a parameter was not used. If there are conflicts with the scripts I edited then discard my minimal changes.